pkg/derivation: move Output to output.go
